+++ Mage/ObservationFilterView.swift
@@ -9,21 +9,14 @@
 import SwiftUI
 
 struct ObservationFilterView: View {
-    @Environment(\.dismiss) var dismiss
-
-    @State private var isFavoriteOn  = false
-    @State private var isImportantOn = false
-    @State private var selectedTime: TimeFilterEnum = .all
-    
-    @State var customTimeFieldValue: Int = UserDefaults.standard.observationTimeFilterNumberKey
-    @State var customTimePickerEnum: TimeUnitWrapper = TimeUnitWrapper(objcValue: UserDefaults.standard.observationTimeFilterUnitKey)
+    @StateObject var observationFilterViewModel = ObservationFilterViewModel()
 
     var body: some View {
         List {
             Section("Filter Types") {
                 HStack {
                     Spacer()
-                    Toggle(isOn: $isFavoriteOn) {
+                    Toggle(isOn: $observationFilterViewModel.isFavoriteOn) {
                         VStack(alignment: .leading) {
                             Text("Favorites")
                             Text("Only show my favorite observations")
@@ -35,7 +28,7 @@ struct ObservationFilterView: View {
                 
                 HStack {
                     Spacer()
-                    Toggle(isOn: $isImportantOn) {
+                    Toggle(isOn: $observationFilterViewModel.isImportantOn) {
                         VStack(alignment: .leading) {
                             Text("Important")
                             Text("Only show observations flagged as important")
@@ -62,11 +55,11 @@ struct ObservationFilterView: View {
                         title: option.title,
                         subTitle: option.subtitle,
                         timeFilter: option,
-                        customTimeFieldValue: $customTimeFieldValue,
-                        customTimePickerEnum: $customTimePickerEnum,
+                        customTimeFieldValue: $observationFilterViewModel.customTimeFieldValue,
+                        customTimePickerEnum: $observationFilterViewModel.customTimePickerEnum,
                         isSelected: Binding(
-                            get: { selectedTime == option },
-                            set: { if $0 { selectedTime = option } }
+                            get: { observationFilterViewModel.selectedTime == option },
+                            set: { if $0 { observationFilterViewModel.selectedTime = option } }
                         )
                     )
                 }
@@ -75,64 +68,15 @@ struct ObservationFilterView: View {
         .navigationTitle("Observation Filter")
         .navigationBarTitleDisplayMode(.inline)
         .toolbarBackground(.visible, for: .navigationBar)
-        .task { loadFromObjC() }
-        .onChange(of: isFavoriteOn)  { saveFavorites($0) }
-        .onChange(of: isImportantOn) { saveImportant($0) }
-        .onChange(of: selectedTime)  { saveTimeFilter($0) }
-        .onChange(of: customTimeFieldValue) { saveCustomTimeFieldValueFilter($0)}
-        .onChange(of: customTimePickerEnum) { saveCustomTimeEnumFilter($0)}
-    }
-
-    private func loadFromObjC() {
-        isFavoriteOn  = Observations.getFavoritesFilter()
-        isImportantOn = Observations.getImportantFilter()
-        selectedTime  = TimeFilterEnum(objc: TimeFilter.getObservationTimeFilter())
-    }
-
-    private func saveFavorites(_ newValue: Bool) {
-        if Observations.getFavoritesFilter() != newValue {
-            Observations.setFavoritesFilter(newValue)
-        }
-    }
-
-    private func saveImportant(_ newValue: Bool) {
-        if Observations.getImportantFilter() != newValue {
-            Observations.setImportantFilter(newValue)
-        }
-    }
-
-    private func saveTimeFilter(_ newValue: TimeFilterEnum) {
-        if TimeFilter.getObservationTimeFilter() != newValue.objc {
-            TimeFilter.setObservation(newValue.objc)
-        }
-    }
-    
-    private func saveCustomTimeFieldValueFilter(_ newValue: Int) {
-        if TimeFilter.getObservationCustomTimeFilterNumber() != newValue {
-            TimeFilter.setObservationCustomTimeFilterNumber(newValue)
-        }
-    }
-    
-    private func saveCustomTimeEnumFilter(_ newValue: TimeUnitWrapper) {
-        if TimeFilter.getObservationCustomTimeFilterUnit() != newValue.objcValue {
-            TimeFilter.setObservationCustomTimeFilterUnit(newValue.objcValue)
-        }
-    }
-
-    // This is for a future method this
-    private func notifyObservationFiltersChanged() {
-        NotificationCenter.default.post(name: .ObservationFiltersChanged, object: nil)
-    }
-
-    private func applyFilterAndDismiss() {
-        saveTimeFilter(selectedTime)
-        saveImportant(isImportantOn)
-        saveFavorites(isFavoriteOn)
-        notifyObservationFiltersChanged()
-        dismiss()
+        .task { observationFilterViewModel.loadFromObjC() }
+        .onChange(of: observationFilterViewModel.isFavoriteOn)  { observationFilterViewModel.saveFavorites($0) }
+        .onChange(of: observationFilterViewModel.isImportantOn) { observationFilterViewModel.saveImportant($0) }
+        .onChange(of: observationFilterViewModel.selectedTime)  { observationFilterViewModel.saveTimeFilter($0) }
+        .onChange(of: observationFilterViewModel.customTimeFieldValue) { observationFilterViewModel.saveCustomTimeFieldValueFilter($0)}
+        .onChange(of: observationFilterViewModel.customTimePickerEnum) { observationFilterViewModel.saveCustomTimeEnumFilter($0)}
     }
 }
 
-#Preview {
-    ObservationFilterView()
-}
+//#Preview {
+//    ObservationFilterView()
+//}
